requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.1
ffmpeg-python==0.2.0
filelock==3.10.7
future==0.18.3
idna==3.4
Jinja2==3.1.2
lit==16.0.0
llvmlite<0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba<0.57.0
numpy==1.23.5
torch
torchaudio
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
transformers==4.35.2
openai-whisper
pytube==15.0.0
yt-dlp==2023.10.13
opencv-python==4.8.1.78
pydantic==2.4.2
python-jose==3.3.0
opencv-python
pytesseract
googletrans